+++ libcext/cext.h
@@ -10,24 +10,24 @@
 #endif
 
 /* array.c */
-void ** cext_array_attach(void **array, void *p, size_t psize, size_t *size);
-void cext_array_detach(void **array, void *p, size_t *size);
+void ** cext_array_attach(void **array, void *p, unsigned int psize, unsigned int *size);
+void cext_array_detach(void **array, void *p, unsigned int *size);
 
 /* emallocz.c */
-void *cext_emallocz(size_t size);
+void *cext_emallocz(unsigned int size);
 
 /* estrdup.c */
 char *cext_estrdup(const char *s);
 
 /* strlcat.c */
-size_t cext_strlcat(char *dst, const char *src, size_t siz);
+unsigned int cext_strlcat(char *dst, const char *src, unsigned int siz);
 
 /* strlcpy.c */
-size_t cext_strlcpy(char *dst, const char *src, size_t siz);
+unsigned int cext_strlcpy(char *dst, const char *src, unsigned int siz);
 
 /* strtonum.c */
 long long cext_strtonum(const char *numstr, long long minval,
                         long long maxval, const char **errstrp);
 
 /* tokenize.c */
-size_t cext_tokenize(char **result, size_t reslen, char *str, char delim);
+unsigned int cext_tokenize(char **result, unsigned int reslen, char *str, char delim);
