+++ libcext/strlcat.c
@@ -26,13 +26,13 @@
  * Returns strlen(src) + MIN(siz, strlen(initial dst)).
  * If retval >= siz, truncation occurred.
  */
-size_t
-cext_strlcat(char *dst, const char *src, size_t siz)
+unsigned int
+cext_strlcat(char *dst, const char *src, unsigned int siz)
 {
     register char *d = dst;
     register const char *s = src;
-    register size_t n = siz;
-    size_t dlen;
+    register unsigned int n = siz;
+    unsigned int dlen;
 
     /* Find the end of dst and adjust bytes left but don't go past end */
     while(n-- != 0 && *d != 0)
